backend/requirements.txt
openai
django
djangorestframework
django-cors-headers
gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib